Restructure project to enhance testing and deployment
