+++ .gitignore
@@ -104,5 +104,6 @@ venv.bak/
 .mypy_cache/
 
 # Project-specific stuff
+.nox/
 .db/
 .idea/
