+++ api/__init__.py
@@ -1,0 +1,5 @@
+from api.config import Config
+from api.main import create_app
+
+
+app = create_app(Config())
