--- requirements.txt
@@ -1,19 +1,0 @@
-aiofiles==0.4.0
-asyncpg==0.20.0
-Click==7.0
-databases==0.2.6
-fastapi==0.44.0
-gunicorn==19.9.0
-h11==0.8.1
-httptools==0.0.13
-migri==0.1.3
-orm==0.1.5
-pydantic==1.2
-SQLAlchemy==1.3.11
-sqlparse==0.3.0
-starlette==0.12.9
-typesystem==0.2.4
-uvicorn==0.10.8
-uvloop==0.14.0
-websockets==8.1
-https://github.com/RonquilloAeon/migri/archive/master.zip
