--- src/start.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-exec gunicorn -w 2 -k uvicorn.workers.UvicornWorker main:app -b 0.0.0.0 "$@"
