+++ start.dev.sh
@@ -4,4 +4,4 @@ while !</dev/tcp/db/5432; do
 sleep 10
 done
 
-exec uvicorn --host 0.0.0.0 main:app "$@"
+exec uvicorn --host 0.0.0.0 api:app "$@"
