+++ start.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+exec gunicorn -w 2 -k uvicorn.workers.UvicornWorker api:app -b 0.0.0.0 "$@"
