+++ test-requirements.txt
@@ -1,0 +1,8 @@
+-e .
+asyncpg==0.14.0
+freezegun==0.3.12
+https://github.com/RonquilloAeon/migri/archive/master.zip
+pytest==5.2.0
+pytest-asyncio==0.10.0
+pytest-cov
+requests==2.22.0
